+++ src/tests/isospin.cc
@@ -43,7 +43,9 @@ static ScatterAction *set_up_action(const ParticleData &proj,
   act = new ScatterAction(proj, targ, 0.);
   CrossSections xs(act->incoming_particles(), act->sqrt_s(),
                    std::make_pair(FourVector(), FourVector()));
-  proc_list = xs.two_to_two(Test::all_reactions_included());
+  proc_list =
+      xs.two_to_two(Test::all_reactions_included(),
+                    InputKeys::collTerm_stringTrans_KNOffset.default_value());
   //   act->add_processes(proc_list);
 
   std::printf("%s+ %s, sqrt(s) = %f GeV, sigma = %f mb, %lu Channels \n",